:building_construction: moved update_game_tick to an other file
