+++ Network/Makefile_Network.mk
@@ -31,6 +31,7 @@ SRC		= 		main.c								\
 				Game/launch_broadcast.c				\
 				Game/object_manipulation.c			\
 				Game/incantation.c					\
+				Game/game_tick.c					\
 
 TRUE_SRC 	= 	$(patsubst %,src/%, $(SRC))
