+++ Network/include/Server/server.h
@@ -153,6 +153,12 @@ in_game_t *init_game(info_game_t info_game);
 void update_players(server_t *server);
 
 /**
+ * @brief Updates the game state.
+ * @param server The server containing the game state to update.
+ */
+void game_tick(server_t *server);
+
+/**
  * @brief sets the number of ticks to wait for
  * the client to execute its next command.
  * @param client The client to set the ticks for.
